Apply port speed again after disabling port AN
Problem and Steps:
Changing the port speed while port AN is enabled will not take effect
after port AN is disabled.

Current behavior/Root Cause:
In PortsOrch, the port speed configuration will not be set to SAI when
port AN is enabled.

Solution:
Apply port speed again after disabling port AN.